requirements.txt
Flask_Bcrypt==0.7.1
WTForms==2.2.1
Flask==1.0.2
itsdangerous==0.24
Flask_Mail==0.9.1
Flask_WTF==0.14.2
Flask_SQLAlchemy==2.3.2
Flask_Login==0.4.1
Pillow==5.2.0
secrets==1.0.2
